LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-import-scope-pseudo.tentative-expected.txt

Inside
Inside, but should not match
Outside

FAIL The :scope pseudo-class works in imported stylesheets assert_equals: expected "1" but got ""
